requirements.txt
Pillow~=7.2.0
setuptools~=49.2.0
eyeD3~=0.9.5